Make sure passphrase was not required for ssh keys
Recent issue came up in which it would be required to use a passphrase for ssh key by default. Instead of having it in use when no password is available, when password (aka "passphrase") and ssh key are available, it would treat the password as a passphrase for the ssh key.